gmail-etl/requirements.txt
# EROS Gmail ETL Requirements
# Core Google Cloud SDKs
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-cloud-secret-manager==2.16.4
google-cloud-bigquery==3.13.0
google-cloud-storage==2.10.0

# Data processing
pandas==2.1.3
numpy==1.25.2
pyarrow==14.0.1

# Excel processing
openpyxl==3.1.2

# Web scraping and downloading
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Additional dependencies
python-dateutil==2.8.2
pytz==2023.3
urllib3==2.0.7